Changed routing for popup and details
