+++ src/components/burger-ingredients/burger-ingredients.js
@@ -3,9 +3,11 @@ import styles from './burger-ingredients.module.css';
 import { Tab } from '@ya.praktikum/react-developer-burger-ui-components'
 import IngredientCard from '../ingredient-card/ingredient-card'
 import { useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
 
 export default function BurgerIngredients() {
+
+  let location = useLocation();
 
   const data = useSelector(store => store.ingredientsReducer.listAllIngredients.data);
   
@@ -70,6 +72,7 @@ export default function BurgerIngredients() {
                         key={card._id}
                         to={{
                           pathname: `/ingredients/${card._id}`,
+                          state: { ingredientModal: location }
                         }}
                       >
                         <IngredientCard key={card._id} card={card} />
